ai_sql_optimizer/app/requirements.txt
fastapi
uvicorn[standard]
requests
SQLAlchemy
psycopg2-binary
python-dotenv
google-cloud-bigquery
pydantic
